pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowstate"
version = "0.1.0"
description = "DJ decision support system - real-time track recommendations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "mutagen>=1.47",
    "rich>=13.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
web = [
    "flask>=3.0",
]

[project.scripts]
flowstate = "flowstate.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/flowstate"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
